feat: remove the old UI of orchestrator and cleanup app and http packages
