--- go/vt/vtorc/app/http.go
@@ -1,170 +1,0 @@
-/*
-   Copyright 2014 Outbrain Inc.
-
-   Licensed under the Apache License, Version 2.0 (the "License");
-   you may not use this file except in compliance with the License.
-   You may obtain a copy of the License at
-
-       http://www.apache.org/licenses/LICENSE-2.0
-
-   Unless required by applicable law or agreed to in writing, software
-   distributed under the License is distributed on an "AS IS" BASIS,
-   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   See the License for the specific language governing permissions and
-   limitations under the License.
-*/
-
-package app
-
-import (
-	"io/ioutil"
-	golog "log"
-	"net"
-	nethttp "net/http"
-	"path"
-	"strings"
-	"time"
-
-	"github.com/spf13/pflag"
-
-	"vitess.io/vitess/go/vt/log"
-
-	"vitess.io/vitess/go/vt/vtorc/collection"
-	"vitess.io/vitess/go/vt/vtorc/config"
-	"vitess.io/vitess/go/vt/vtorc/http"
-	"vitess.io/vitess/go/vt/vtorc/inst"
-	"vitess.io/vitess/go/vt/vtorc/logic"
-	"vitess.io/vitess/go/vt/vtorc/process"
-	"vitess.io/vitess/go/vt/vtorc/ssl"
-
-	"github.com/go-martini/martini"
-	"github.com/martini-contrib/auth"
-	"github.com/martini-contrib/gzip"
-	"github.com/martini-contrib/render"
-)
-
-const discoveryMetricsName = "DISCOVERY_METRICS"
-
-var webDir = "web/vtorc"
-
-var sslPEMPassword []byte
-var discoveryMetrics *collection.Collection
-
-// RegisterFlags registers the flags required by VTOrc
-func RegisterFlags(fs *pflag.FlagSet) {
-	fs.StringVar(&webDir, "orc_web_dir", webDir, "VTOrc http file location")
-}
-
-// HTTP starts serving
-func HTTP(continuousDiscovery bool) {
-	promptForSSLPasswords()
-	process.ContinuousRegistration(string(process.VTOrcExecutionHTTPMode), "")
-
-	martini.Env = martini.Prod
-	standardHTTP(continuousDiscovery)
-}
-
-// Iterate over the private keys and get passwords for them
-// Don't prompt for a password a second time if the files are the same
-func promptForSSLPasswords() {
-	if ssl.IsEncryptedPEM(config.Config.SSLPrivateKeyFile) {
-		sslPEMPassword = ssl.GetPEMPassword(config.Config.SSLPrivateKeyFile)
-	}
-}
-
-// standardHTTP starts serving HTTP or HTTPS (api/web) requests, to be used by normal clients
-func standardHTTP(continuousDiscovery bool) {
-	m := martini.Classic()
-	// Make martini silent by setting its logger to a discard endpoint
-	m.Logger(golog.New(ioutil.Discard, "", 0))
-
-	switch strings.ToLower(config.Config.AuthenticationMethod) {
-	case "basic":
-		{
-			if config.Config.HTTPAuthUser == "" {
-				// Still allowed; may be disallowed in future versions
-				log.Warning("AuthenticationMethod is configured as 'basic' but HTTPAuthUser undefined. Running without authentication.")
-			}
-			m.Use(auth.Basic(config.Config.HTTPAuthUser, config.Config.HTTPAuthPassword))
-		}
-	case "multi":
-		{
-			if config.Config.HTTPAuthUser == "" {
-				// Still allowed; may be disallowed in future versions
-				log.Fatal("AuthenticationMethod is configured as 'multi' but HTTPAuthUser undefined")
-			}
-
-			m.Use(auth.BasicFunc(func(username, password string) bool {
-				if username == "readonly" {
-					// Will be treated as "read-only"
-					return true
-				}
-				return auth.SecureCompare(username, config.Config.HTTPAuthUser) && auth.SecureCompare(password, config.Config.HTTPAuthPassword)
-			}))
-		}
-	default:
-		{
-			// We inject a dummy User object because we have function signatures with User argument in api.go
-			m.Map(auth.User(""))
-		}
-	}
-
-	m.Use(gzip.All())
-	// Render html templates from templates directory
-	m.Use(render.Renderer(render.Options{
-		Directory:       webDir,
-		Layout:          "templates/layout",
-		HTMLContentType: "text/html",
-	}))
-	m.Use(martini.Static(path.Join(webDir, "public"), martini.StaticOptions{Prefix: config.Config.URLPrefix}))
-	if config.Config.UseMutualTLS {
-		m.Use(ssl.VerifyOUs(config.Config.SSLValidOUs))
-	}
-
-	inst.SetMaintenanceOwner(process.ThisHostname)
-
-	if continuousDiscovery {
-		// start to expire metric collection info
-		discoveryMetrics = collection.CreateOrReturnCollection(discoveryMetricsName)
-		discoveryMetrics.SetExpirePeriod(time.Duration(config.Config.DiscoveryCollectionRetentionSeconds) * time.Second)
-
-		log.Info("Starting Discovery")
-		go logic.ContinuousDiscovery()
-	}
-
-	log.Info("Registering endpoints")
-	http.HTTPWeb.URLPrefix = config.Config.URLPrefix
-	http.HTTPWeb.RegisterRequests(m)
-
-	// Serve
-	if config.Config.ListenSocket != "" {
-		log.Infof("Starting HTTP listener on unix socket %v", config.Config.ListenSocket)
-		unixListener, err := net.Listen("unix", config.Config.ListenSocket)
-		if err != nil {
-			log.Fatal(err)
-		}
-		defer unixListener.Close()
-		if err := nethttp.Serve(unixListener, m); err != nil {
-			log.Fatal(err)
-		}
-	} else if config.Config.UseSSL {
-		log.Info("Starting HTTPS listener")
-		tlsConfig, err := ssl.NewTLSConfig(config.Config.SSLCAFile, config.Config.UseMutualTLS)
-		if err != nil {
-			log.Fatal(err)
-		}
-		tlsConfig.InsecureSkipVerify = config.Config.SSLSkipVerify
-		if err = ssl.AppendKeyPairWithPassword(tlsConfig, config.Config.SSLCertFile, config.Config.SSLPrivateKeyFile, sslPEMPassword); err != nil {
-			log.Fatal(err)
-		}
-		if err = ssl.ListenAndServeTLS(config.Config.ListenAddress, m, tlsConfig); err != nil {
-			log.Fatal(err)
-		}
-	} else {
-		log.Infof("Starting HTTP listener on %+v", config.Config.ListenAddress)
-		if err := nethttp.ListenAndServe(config.Config.ListenAddress, m); err != nil {
-			log.Fatal(err)
-		}
-	}
-	log.Info("Web server closed")
-}
