--- go/vt/vtorc/http/api.go
@@ -1,66 +1,0 @@
-/*
-   Copyright 2014 Outbrain Inc.
-
-   Licensed under the Apache License, Version 2.0 (the "License");
-   you may not use this file except in compliance with the License.
-   You may obtain a copy of the License at
-
-       http://www.apache.org/licenses/LICENSE-2.0
-
-   Unless required by applicable law or agreed to in writing, software
-   distributed under the License is distributed on an "AS IS" BASIS,
-   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   See the License for the specific language governing permissions and
-   limitations under the License.
-*/
-
-package http
-
-import (
-	"encoding/json"
-	"net/http"
-)
-
-// APIResponseCode is an OK/ERROR response code
-type APIResponseCode int
-
-const (
-	ERROR APIResponseCode = iota
-	OK
-)
-
-func (apiResponseCode *APIResponseCode) MarshalJSON() ([]byte, error) {
-	return json.Marshal(apiResponseCode.String())
-}
-
-func (apiResponseCode *APIResponseCode) String() string {
-	switch *apiResponseCode {
-	case ERROR:
-		return "ERROR"
-	case OK:
-		return "OK"
-	}
-	return "unknown"
-}
-
-// HTTPStatus returns the respective HTTP status for this response
-func (apiResponseCode *APIResponseCode) HTTPStatus() int {
-	switch *apiResponseCode {
-	case ERROR:
-		return http.StatusInternalServerError
-	case OK:
-		return http.StatusOK
-	}
-	return http.StatusNotImplemented
-}
-
-// APIResponse is a response returned as JSON to various requests.
-type APIResponse struct {
-	Code    APIResponseCode
-	Message string
-	Details any
-}
-
-type API struct {
-	URLPrefix string
-}
